+++ lib/gen/assets.gen.dart
@@ -1,0 +1,86 @@
+//          --TURN_GEN--
+//           v0.6.3 (assets)
+//  *************************************
+//         GENERATED CODE
+//  *************************************
+
+// coverage:ignore-file
+// ignore_for_file: type=lint
+// ignore_for_file: directives_ordering,unnecessary_import,implicit_dynamic_list_literal,deprecated_member_use
+
+class AppAssets {
+ 
+  /// * Size:	113.2 KB
+  /// * File path: _/assets/fonts/TTCommons-SemiBold.ttf
+  static const String fontsTTCommonsSemiBold =
+      '/assets/fonts/TTCommons-SemiBold.ttf';
+
+  /// * Size:	59.1 KB
+  /// * File path: _/assets/image/logo.png
+  static const String imageLogo = '/assets/image/logo.png';
+
+  /// * Size:	12.9 KB
+  /// * File path: _/assets/image/logo_card_1.png
+  static const String imageLogoCard1 = '/assets/image/logo_card_1.png';
+
+ 
+
+  /// * Size:	26.0 KB
+  /// * File path: _/assets/lottie/load_page.json
+  static const String lottieLoadPage = '/assets/lottie/load_page.json';
+
+  /// * Size:	12.4 KB
+  /// * File path: _/assets/svg/card_photo.svg
+  static const String svgCardPhoto = '/assets/svg/card_photo.svg';
+
+  /// * Size:	6.8 KB
+  /// * File path: _/assets/svg/connectivity_problem.svg
+  static const String svgConnectivityProblem =
+      '/assets/svg/connectivity_problem.svg';
+
+  /// * Size:	1.7 KB
+  /// * File path: _/assets/svg/doc.svg
+  static const String svgDoc = '/assets/svg/doc.svg';
+
+  /// * Size:	744.0 B
+  /// * File path: _/assets/svg/valid_error.svg
+  static const String svgValidError = '/assets/svg/valid_error.svg';
+
+  /// List of PNG assets
+  static const List<String> valuesPNG = [
+  
+    imageLogo,
+    imageLogoCard1,
+  
+  ];
+
+  /// List of TTF assets
+  static const List<String> valuesTTF = [
+  
+    fontsTTCommonsSemiBold
+  ];
+
+  /// List of JSON assets
+  static const List<String> valuesJSON = [lottieLoadBtn, lottieLoadPage];
+
+  /// List of SVG assets
+  static const List<String> valuesSVG = [
+    svgCardPhoto,
+    svgConnectivityProblem,
+    svgDoc,
+    svgValidError
+  ];
+
+  /// List of all assets
+  static const List<String> valuesAll = [
+  
+    fontsTTCommonsSemiBold,
+    imageLogo,
+    imageLogoCard1,
+    lottieLoadPage,
+    svgCardPhoto,
+    svgConnectivityProblem,
+    svgDoc,
+    svgValidError
+  ];
+}