+++ lib/src/assets_gen/assets_gen.dart
@@ -1,4 +1,5 @@
 export 'enum_type_assets.dart';
-export 'model/model.dart';
+export 'model/asset_item.dart';
+export 'ru_to_lat/alpabet.dart';
 export 'ru_to_lat/ru_to_lat.dart';
 export 'run_assets.dart';