+++ lib/src/utils.dart
@@ -2,7 +2,9 @@ import 'dart:io';
 
 extension StringCasingExtension on String {
   String toCapitalized() =>
-      length > 0 ? '${this[0].toUpperCase()}${substring(1).toLowerCase()}' : '';
+      length > 0 ? '${this[0].toUpperCase()}${substring(1)}' : '';
+  String toNotTitle() =>
+      length > 0 ? '${this[0].toLowerCase()}${substring(1)}' : '';
   String toTitleCase() => replaceAll(RegExp(' +'), ' ')
       .split(' ')
       .map((str) => str.toCapitalized())